server/requirements.txt
boto3
flask
python-dotenv
requests
pillow
matplotlib
opencv-python
ipython
pip-api
